feat(app): add crud form quote

+++ src/Controller/QuoteController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Quote;
 use App\Form\QuoteType;
 use App\Repository\QuoteRepository;
+use App\Services\QuoteService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -15,6 +16,7 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class QuoteController extends AbstractController
 {
+
     /**
      * @Route("/", name="quote_index", methods={"GET"})
      */
@@ -31,9 +33,7 @@ class QuoteController extends AbstractController
     public function new(Request $request): Response
     {
         $quote = new Quote();
-
         $form = $this->createForm(QuoteType::class, $quote);
-
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -41,25 +41,57 @@ class QuoteController extends AbstractController
             $entityManager->persist($quote);
             $entityManager->flush();
 
-            return $this->redirectToRoute('quotes');
+            return $this->redirectToRoute('quote_index');
         }
 
-        return $this->render('quote/index.html.twig', [
+        return $this->render('quote/new.html.twig', [
             'quote' => $quote,
             'form' => $form->createView(),
         ]);
     }
 
     /**
-     * @Route("/quote_delete/{id}", name="quote_delete")
+     * @Route("/{id}", name="quote_show", methods={"GET"})
      */
-    public function delete(Quote $quote): Response
+
+    public function show(QuoteService $quote,$id): Response
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($quote);
-        $entityManager->flush();
+        return $this->render('quote/show.html.twig', [
+            'quote' => $quote->getQuote($id),
+        ]);
+    }
+
+    /**
+     * @Route("/{id}/edit", name="quote_edit", methods={"GET","POST"})
+     */
+    public function edit(Request $request, Quote $quote): Response
+    {
+        $form = $this->createForm(QuoteType::class, $quote);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->getDoctrine()->getManager()->flush();
+
+            return $this->redirectToRoute('quotes');
+        }
+
+        return $this->render('quote/edit.html.twig', [
+            'quote' => $quote,
+            'form' => $form->createView(),
+        ]);
+    }
+
+    /**
+     * @Route("/{id}", name="quote_delete", methods={"DELETE"})
+     */
+    public function delete(Request $request, Quote $quote): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$quote->getId(), $request->request->get('_token'))) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($quote);
+            $entityManager->flush();
+        }
 
         return $this->redirectToRoute('quotes');
     }
-
 }
